feat: 외부 API 호출 계층 컨트롤러에서 서비스로 변경
