+++ src/main/java/roomescape/reservation/controller/ReservationApiController.java
@@ -1,8 +1,6 @@
 package roomescape.reservation.controller;
 
 import java.net.URI;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.List;
 
 import jakarta.validation.Valid;
@@ -17,9 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.auth.Login;
-import roomescape.client.PaymentClient;
 import roomescape.member.dto.LoginMemberInToken;
-import roomescape.reservation.dto.request.PaymentRequest;
 import roomescape.reservation.dto.request.ReservationCreateRequest;
 import roomescape.reservation.dto.request.ReservationSearchRequest;
 import roomescape.reservation.dto.response.MyReservationResponse;
@@ -29,14 +25,10 @@ import roomescape.reservation.service.ReservationService;
 
 @RestController
 public class ReservationApiController {
-    private static final String WIDGET_SECRET_KEY = "test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6";
-
     private final ReservationService reservationService;
-    private final PaymentClient paymentClient;
 
-    public ReservationApiController(ReservationService reservationService, PaymentClient paymentClient) {
+    public ReservationApiController(ReservationService reservationService) {
         this.reservationService = reservationService;
-        this.paymentClient = paymentClient;
     }
 
     @GetMapping("/reservations")
@@ -60,19 +52,10 @@ public class ReservationApiController {
             @Valid @RequestBody ReservationCreateRequest reservationCreateRequest,
             @Login LoginMemberInToken loginMemberInToken
     ) {
-        paymentClient.paymentReservation(getAuthorizations(),
-                PaymentRequest.toRequest(reservationCreateRequest)).getBody();
-
         Long id = reservationService.save(reservationCreateRequest, loginMemberInToken);
         ReservationResponse reservationResponse = reservationService.findById(id);
 
         return ResponseEntity.created(URI.create("/reservations/" + id)).body(reservationResponse);
-    }
-
-    private String getAuthorizations() {
-        Base64.Encoder encoder = Base64.getEncoder();
-        byte[] encodedBytes = encoder.encode((WIDGET_SECRET_KEY + ":").getBytes(StandardCharsets.UTF_8));
-        return "Basic " + new String(encodedBytes);
     }
 
     @GetMapping("/reservations/waiting")
