+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,8 +1,13 @@
 package roomescape.reservation.service;
 
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import roomescape.client.PaymentClient;
 import roomescape.member.domain.Member;
 import roomescape.member.dto.LoginMemberInToken;
 import roomescape.member.repository.MemberRepository;
@@ -11,6 +16,7 @@ import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.Waitings;
+import roomescape.reservation.dto.request.PaymentRequest;
 import roomescape.reservation.dto.request.ReservationCreateRequest;
 import roomescape.reservation.dto.request.ReservationSearchRequest;
 import roomescape.reservation.dto.response.MyReservationResponse;
@@ -23,35 +29,49 @@ import roomescape.reservation.repository.ThemeRepository;
 @Service
 @Transactional
 public class ReservationService {
+    private static final String WIDGET_SECRET_KEY = "test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6";
 
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
+    private final PaymentClient paymentClient;
+
 
     public ReservationService(
             ReservationRepository reservationRepository,
             ReservationTimeRepository reservationTimeRepository,
             ThemeRepository themeRepository,
-            MemberRepository memberRepository
+            MemberRepository memberRepository,
+            PaymentClient paymentClient
     ) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
+        this.paymentClient = paymentClient;
     }
 
     public Long save(ReservationCreateRequest reservationCreateRequest, LoginMemberInToken loginMemberInToken) {
+        paymentClient.paymentReservation(getAuthorizations(), PaymentRequest.toRequest(reservationCreateRequest));
+
         if (reservationRepository.existsByDateAndReservationTimeIdAndThemeId(reservationCreateRequest.date(),
                 reservationCreateRequest.timeId(), reservationCreateRequest.themeId())) {
             Reservation reservation = getValidatedReservation(reservationCreateRequest, loginMemberInToken,
                     Status.WAITING);
             return reservationRepository.save(reservation).getId();
         }
-
         Reservation reservation = getValidatedReservation(reservationCreateRequest, loginMemberInToken, Status.SUCCESS);
         return reservationRepository.save(reservation).getId();
     }
+
+
+    private String getAuthorizations() {
+        Base64.Encoder encoder = Base64.getEncoder();
+        byte[] encodedBytes = encoder.encode((WIDGET_SECRET_KEY + ":").getBytes(StandardCharsets.UTF_8));
+        return "Basic " + new String(encodedBytes);
+    }
+
 
     private Reservation getValidatedReservation(ReservationCreateRequest reservationCreateRequest,
                                                 LoginMemberInToken loginMemberInToken,
@@ -62,13 +82,12 @@ public class ReservationService {
         Theme theme = themeRepository.findById(reservationCreateRequest.themeId())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마입니다."));
 
-        Member member = getValidatedMemberByRole(reservationCreateRequest, loginMemberInToken);
+        Member member = getValidatedMemberByRole(loginMemberInToken);
 
         return reservationCreateRequest.toReservation(member, theme, reservationTime, status);
     }
 
-    private Member getValidatedMemberByRole(ReservationCreateRequest reservationCreateRequest,
-                                            LoginMemberInToken loginMemberInToken) {
+    private Member getValidatedMemberByRole(LoginMemberInToken loginMemberInToken) {
         return memberRepository.findById(loginMemberInToken.id())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
     }
@@ -87,8 +106,10 @@ public class ReservationService {
     }
 
     public List<ReservationResponse> findAllBySearch(ReservationSearchRequest reservationSearchRequest) {
-        Theme theme = themeRepository.findById(reservationSearchRequest.themeId()).get();
-        Member member = memberRepository.findById(reservationSearchRequest.memberId()).get();
+        Theme theme = themeRepository.findById(reservationSearchRequest.themeId())
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마 정보 입니다."));
+        Member member = memberRepository.findById(reservationSearchRequest.memberId())
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원 정보입니다."));
         return reservationRepository.findAllByMemberAndThemeAndDateBetween(member, theme,
                         reservationSearchRequest.dateFrom(), reservationSearchRequest.dateTo()).stream()
                 .map(ReservationResponse::toResponse)
@@ -107,7 +128,8 @@ public class ReservationService {
 
     @Transactional
     public void delete(Long id) {
-        Reservation reservation = reservationRepository.findById(id).get();
+        Reservation reservation = reservationRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약 정보입니다."));
         reservationRepository.deleteById(id);
 
         Waitings waitings = new Waitings(reservationRepository.findAllByDateAndReservationTimeIdAndThemeIdAndStatus(
